requirements.txt
numpy
prettytable
paramiko
pyaml
botocore
boto3
